Large update (MapMaker, LayerSizeEditor, tileMap alterations, cleanup with RenderManager, etc)
Implemented the first go at the MapMaker state, so far a couple buttons
have been made and let the user at least create a map of defaultly
initialized tiles and resize the map (small but still left where pins
can be stretched 'outside' of legal grid area. This is entirely a grid
coordinate issue with negatives and all.
- Some clean up in RenderManager, WorldManager, tile map data structure,
and SimulationState.
- Added some self made buttons to the resource folder.
- Refactored the naming of several methods.
- Changed the 'defaultInitNullTiles' method to allow for only
initializing null tiles so as not to overwrite already initialized
tiles.
- Added some clean up code for getting rid of entities when deleting
tiles and layers. ( prepForDeletion() )